[SYCL][NVPTX] Add a pass to promote kernel args to constant memory.
This pass operates on SYCL kernels being compiled to CUDA. It rewrites
the signature of the kernel, removing all the arguments, instead the
arguments are used to construct a struct, where each kernel argument is
a struct member. The pass then constructs a global variable of the
struct type (in constant address space) and replaces the use of all the
arguments with the corresponding members. The SYCL runtime is expected
to fill in the struct with the data taken from appropriate kernel
arguments.