[Sema] Wrap SYCL kernel object in union
Wrap the SYCL kernel object in an union.
This works around the clang codegen emiting the object destructor.

User can declare their type as device copyable even
if they do not meet the criterias. In some cases,
this triggers calls to undefined function.

Per SYCL rules, a SYCL kernel object must be device copiable
which normally implies a trivial destructor. Plus the spec doesn't guarantee
that the destructor should get called at all as it allows the functor
to be reused by multiple WI.

This patch exploit that UB by preventing the destructor of being called
by wrapping it into a union. This does nothing to prevent cleanup
of temporaries if they happen.

Signed-off-by: Victor Lomuller <victor@codeplay.com>