[SROA] Add struct "peeler" by working out lower bounds
Allows SROA to compute lower bounds when the main algorithm aborts.
By getting lower bound, we can then extract chunks of a struct.

This is useful when array within a struct is accessed in a loop.
This case would typically bailout due to the dynamic access of the array.
If the access to that array is inbound, we can speculate that
it is safe to move it into its own alloca.

One side that is probably not yet well handled is when the struct
is reinterpreted into an other type and lower bound type traversed.
We should bail out in these cases, but probably don't.

Signed-off-by: Victor Lomuller <victor@codeplay.com>