Introduce an improved address space inference checker
The patch introduce a proof of concept for a new address space
inference checker for SYCL. The approach rely on aggregating type
constraints of a SYCL device program and use simple unification
algorithm in order to efficiently solve them.

In order to triger the inference, simply add
-fsycl-check-address-space-infer to the clang invocation.

There is 2 main part to the checker:

- Unifier: include the unifier and the type abstraction used to
           represent constraints
- ConstraintGather / ConstraintGatherCast: it is a Decl and Stmt
           visitor parsing the SYCL program function per function
           and add constraints to the unifier

Tests are implemented as clang unit and lit test
When running check-sycl, -fsycl-check-address-space-infer is
automatically set. Most of the test pass but there is still about
90 failures.